refactor(launcher): extract connection handlers into reducer/connection.rs (task #182 PR-E)
Final per-domain extraction in the launcher reducer split. Brings
mod.rs to 564 lines — under the 600-line target from task #182.

Extracted:
- enforce_host_only (gate used by every host-mutating arm)
- host_is_running (predicate for OrphanInstance drift detection)
- handle_get_snapshot
- handle_register
- handle_goodbye

All marked pub(super). Dispatcher arms updated to call connection::*.
window.rs's super::host_is_running → super::connection::host_is_running.
tests.rs's host_is_running → super::connection::host_is_running.

Sizes:
- reducer/mod.rs: 765 → 564 lines (-201)
- reducer/connection.rs: NEW 221 lines

## Final per-domain layout (post task #182 PR-A through PR-E)

| File | Lines | Contents |
|---|---|---|
| reducer/mod.rs | 564 | dispatcher (update), Ctx, Command-type-specific wrong-pipe arms |
| reducer/tests.rs | 2765 | all tests |
| reducer/connection.rs | 221 | register, ping, goodbye, gate, host-running |
| reducer/window.rs | 215 | window opened/closed, backend window IDs |
| reducer/pool.rs | 116 | pool window add/remove/promote, drain decision |
| reducer/saga.rs | 66 | panes_reaped, saga_action_failed |

154 launcher tests pass.
